requirements.txt
bedrock-agentcore
boto3
python-dotenv
httpx
